+++ gtfs-validation-lib/src/test/resources/gtfs/agency.txt
@@ -1,0 +1,2 @@
+agency_id,agency_name,agency_url,agency_timezone,agency_lang,agency_phone
+SoundTransit,Sound Transit,http://www.soundtransit.org,America/Los_Angeles,EN,(800) 201-4900
