+++ gtfs-validation-lib/src/test/resources/gtfs/calendar.txt
@@ -1,0 +1,4 @@
+service_id,monday,tuesday,wednesday,thursday,friday,saturday,sunday,start_date,end_date
+SU,0,0,0,0,0,0,1,20130929,20140215
+SA,0,0,0,0,0,1,0,20130929,20140215
+WD,1,1,1,1,1,0,0,20130929,20140215
