+++ gtfs-validation-lib/src/test/resources/gtfs/calendar_dates.txt
@@ -1,0 +1,7 @@
+service_id,date,exception_type
+WD,20131128,2
+SU,20131128,1
+WD,20131225,2
+SU,20131225,1
+WD,20140101,2
+SU,20140101,1
