+++ gtfs-validation-lib/src/test/resources/gtfs/fare_attributes.txt
@@ -1,0 +1,26 @@
+fare_id,price,currency_type,payment_method,transfers,transfer_duration
+13,3.00,USD,1,,7200
+12,2.75,USD,1,,7200
+14,3.25,USD,1,,7200
+15,3.50,USD,1,,7200
+16,3.75,USD,1,,7200
+17,4.00,USD,1,,7200
+18,4.25,USD,1,,7200
+19,4.50,USD,1,,7200
+20,4.75,USD,1,,7200
+21,5.00,USD,1,,7200
+22,5.25,USD,1,,7200
+30,2.50,USD,0,,7200
+34,3.00,USD,0,,7200
+36,3.50,USD,0,,7200
+1,0.00,USD,0,,
+109,1.5,USD,1,,7200
+110,1.5,USD,1,,7200
+111,2.5,USD,1,,7200
+112,2.5,USD,1,,7200
+113,3,USD,1,,7200
+114,3,USD,1,,7200
+218,2,USD,1,,7200
+220,2.25,USD,1,,7200
+221,2.5,USD,1,,7200
+222,2.75,USD,1,,7200
