+++ gtfs-validation-lib/src/test/resources/gtfs/fare_rules.txt
@@ -1,0 +1,80 @@
+fare_id,route_id,origin_id,destination_id,contains_id
+12,,S_ST,S_LW,
+12,,S_ST,S_TD,
+14,,S_ST,S_PU,
+15,,S_ST,S_SU,
+17,,S_ST,S_AU,
+18,,S_ST,S_KE,
+19,,S_ST,S_TK,
+21,,S_ST,S_KS,
+12,,S_LW,S_TD,
+15,,S_LW,S_PU,
+15,,S_LW,S_SU,
+17,,S_LW,S_AU,
+18,,S_LW,S_KE,
+19,,S_LW,S_TK,
+21,,S_LW,S_KS,
+20,,S_TD,S_KS,
+17,,S_TD,S_TK,
+16,,S_TD,S_KE,
+15,,S_TD,S_AU,
+13,,S_TD,S_SU,
+13,,S_TD,S_PU,
+18,,S_PU,S_KS,
+16,,S_PU,S_TK,
+15,,S_PU,S_KE,
+13,,S_PU,S_AU,
+12,,S_PU,S_SU,
+13,,S_PU,S_TD,
+18,,S_SU,S_KS,
+15,,S_SU,S_TK,
+14,,S_SU,S_KE,
+13,,S_SU,S_AU,
+12,,S_SU,S_PU,
+13,,S_SU,S_TD,
+16,,S_AU,S_KS,
+14,,S_AU,S_TK,
+12,,S_AU,S_KE,
+13,,S_AU,S_SU,
+13,,S_AU,S_PU,
+15,,S_AU,S_TD,
+15,,S_KE,S_KS,
+12,,S_KE,S_TK,
+12,,S_KE,S_AU,
+14,,S_KE,S_SU,
+15,,S_KE,S_PU,
+16,,S_KE,S_TD,
+14,,S_TK,S_KS,
+12,,S_TK,S_KE,
+14,,S_TK,S_AU,
+15,,S_TK,S_SU,
+16,,S_TK,S_PU,
+17,,S_TK,S_TD,
+20,,S_KS,S_TK,
+14,,S_KS,S_KE,
+16,,S_KS,S_AU,
+18,,S_KS,S_SU,
+18,,S_KS,S_PU,
+20,,S_KS,S_TD,
+12,,S_EV,S_MU,
+15,,S_EV,S_ED,
+19,,S_EV,S_KS,
+14,,S_MU,S_ED,
+12,,S_MU,S_EV,
+17,,S_MU,S_KS,
+14,,S_ED,S_MU,
+15,,S_ED,S_EV,
+15,,S_ED,S_KS,
+17,,S_KS,S_MU,
+15,,S_KS,S_ED,
+19,,S_KS,S_EV,
+1,,TLZ,TLZ,
+30,,A,A,
+36,,A,B,
+36,,A,C,
+36,,B,A,
+30,,B,B,
+30,,B,C,
+36,,C,A,
+30,,C,B,
+30,,C,C,
