+++ gtfs-validation-lib/src/test/resources/gtfs/frequencies.txt
@@ -1,0 +1,11 @@
+trip_id,start_time,end_time,headway_secs
+TLNBSA,08:00:00,22:00:00,720
+TLSBSA,08:00:00,22:00:00,720
+TLNBSU,10:00:00,18:00:00,1440
+TLSBSU,10:00:00,18:00:00,1440
+TLNBWD,05:30:00,07:00:00,1440
+TLNBWD,07:00:00,20:00:00,720
+TLNBWD,20:00:00,22:00:00,1440
+TLSBWD,05:30:00,07:00:00,1440
+TLSBWD,07:00:00,20:00:00,720
+TLSBWD,20:00:00,22:00:00,1440
