+++ gtfs-validation-lib/src/test/resources/gtfs/routes.txt
@@ -1,0 +1,17 @@
+route_id,route_short_name,route_long_name,route_desc,route_type,route_url
+SNDR_S,,Sounder / South,Sounder South runs between Pierce County and King Street Station in Seattle weekdays with occasional special service on weekends serving special events,2,
+SNDR_N,,Sounder / North,Sounder North runs between Everett Station and King Street Station in Seattle weekdays with occasional special service on weekends serving special events,2,
+AMTK,,Amtrak,Amtrak service supplements Sounder North weekdays.  Please note that Amtrak does not stop in Mukilteo and accepts only passes valid on Sounder service,2,
+TLINK,,Tacoma Link,Tacoma Link provides service through downtown Tacoma 7 days a week,0,
+510N,510,Seattle - Everett,Sound Transit Route,3,
+510S,510,Everett - Seattle,Sound Transit Route,3,
+511N,511,Seattle - Ash Way,Sound Transit Route,3,
+511S,511,Ash Way - Seattle,Sound Transit Route,3,
+512N,512,Seattle - Everett,Sound Transit Route,3,
+512S,512,Everett - Seattle,Sound Transit Route,3,
+513N,513,Seattle - Everett,Sound Transit Route,3,
+513S,513,Everett - Seattle,Sound Transit Route,3,
+532N,532,Bellevue - Everett,Sound Transit Route,3,
+532S,532,Everett - Bellevue,Sound Transit Route,3,
+535N,535,Bellevue - Lynnwood,Sound Transit Route,3,
+535S,535,Lynnwood - Bellevue,Sound Transit Route,3,
